+++ packages/headless/src/app/headless-engine.test.ts
@@ -9,6 +9,12 @@ import {SearchAPIClient} from '../api/search/search-api-client';
 import {AnalyticsClientSendEventHook} from 'coveo.analytics/dist/definitions/client/analytics';
 import pino from 'pino';
 import {NoopPreprocessRequestMiddleware} from '../api/platform-client';
+import {validatePayloadAndThrow} from '../utils/validate-payload';
+import {
+  NoopPostprocessFacetSearchResponseMiddleware,
+  NoopPostprocessQuerySuggestResponseMiddleware,
+  NoopPostprocessSearchResponseMiddleware,
+} from '../api/search/search-api-client-middleware';
 
 describe('headless engine', () => {
   let options: HeadlessOptions<typeof searchAppReducers>;
@@ -25,9 +31,13 @@ describe('headless engine', () => {
           logger,
           renewAccessToken: async () => '',
           preprocessRequest: NoopPreprocessRequestMiddleware,
+          postprocessSearchResponseMiddleware: NoopPostprocessSearchResponseMiddleware,
+          postprocessQuerySuggestResponseMiddleware: NoopPostprocessQuerySuggestResponseMiddleware,
+          postprocessFacetSearchResponseMiddleware: NoopPostprocessFacetSearchResponseMiddleware,
         }),
         analyticsClientMiddleware: {} as AnalyticsClientSendEventHook,
         logger: logger,
+        validatePayload: validatePayloadAndThrow,
       },
     });
     jest.spyOn(store, 'dispatch');