+++ packages/headless/src/controllers/result-list/headless-result-list.ts
@@ -41,6 +41,7 @@ export function buildResultList(
   const {dispatch} = engine;
 
   const options = validateOptions(
+    engine,
     optionsSchema,
     props?.options,
     buildResultList.name