+++ packages/headless/src/features/facets/range-facets/date-facet-set/date-facet-controller-actions.ts
@@ -7,7 +7,6 @@ import {
 } from '../../../../state/state-sections';
 import {executeToggleRangeFacetSelect} from '../generic/range-facet-controller-actions';
 import {toggleSelectDateFacetValue} from './date-facet-actions';
-import {validatePayloadSchema} from '../../../../utils/validate-payload';
 import {facetIdDefinition} from '../../generic/facet-actions-validation';
 import {RecordValue} from '@coveo/bueno';
 import {dateFacetValueDefinition} from '../generic/range-facet-validate-payload';
@@ -24,14 +23,17 @@ export const executeToggleDateFacetSelect = createAsyncThunk<
     selection: DateFacetValue;
   },
   AsyncThunkSearchOptions<ConfigurationSection & DateFacetSection>
->('dateFacet/executeToggleSelect', ({facetId, selection}, {dispatch}) => {
-  validatePayloadSchema(
-    {facetId, selection},
-    {
-      facetId: facetIdDefinition,
-      selection: new RecordValue({values: dateFacetValueDefinition}),
-    }
-  );
-  dispatch(toggleSelectDateFacetValue({facetId, selection}));
-  dispatch(executeToggleRangeFacetSelect({facetId, selection}));
-});
+>(
+  'dateFacet/executeToggleSelect',
+  ({facetId, selection}, {dispatch, extra: {validatePayload}}) => {
+    validatePayload(
+      {facetId, selection},
+      {
+        facetId: facetIdDefinition,
+        selection: new RecordValue({values: dateFacetValueDefinition}),
+      }
+    );
+    dispatch(toggleSelectDateFacetValue({facetId, selection}));
+    dispatch(executeToggleRangeFacetSelect({facetId, selection}));
+  }
+);