+++ packages/headless/src/features/recommendation/recommendation-actions.ts
@@ -16,7 +16,7 @@ import {
   RecommendationSection,
   SearchHubSection,
 } from '../../state/state-sections';
-import {validatePayloadSchema} from '../../utils/validate-payload';
+import {validatePayload} from '../../utils/validate-payload';
 import {StringValue} from '@coveo/bueno';
 import {logRecommendationUpdate} from './recommendation-analytics-actions';
 import {SearchAction} from '../analytics/analytics-utils';
@@ -44,7 +44,7 @@ export interface GetRecommendationsThunkReturn {
 export const setRecommendationId = createAction(
   'recommendation/set',
   (payload: {id: string}) =>
-    validatePayloadSchema(payload, {
+    validatePayload(payload, {
       id: new StringValue({required: true, emptyAllowed: false}),
     })
 );