+++ packages/headless/src/features/search-parameters/search-parameter-schema.ts
@@ -19,4 +19,5 @@ export const searchParametersDefinition: SchemaDefinition<Required<
   sortCriteria: new StringValue(),
   f: new RecordValue(),
   cf: new RecordValue(),
+  debug: new BooleanValue(),
 };