+++ packages/headless/src/features/search-parameters/search-parameter-serializer.ts
@@ -115,6 +115,7 @@ function isValidKey(key: string): key is keyof SearchParameters {
     sortCriteria: true,
     f: true,
     cf: true,
+    debug: true,
   };
 
   return key in supportedParameters;
@@ -126,6 +127,10 @@ function cast<K extends keyof SearchParameters>(
   const [key, value] = pair;
 
   if (key === 'enableQuerySyntax') {
+    return [key, value === 'true'];
+  }
+
+  if (key === 'debug') {
     return [key, value === 'true'];
   }